General Upgrades
* Upgrade Maven Wrapper and Maven
* Upgrade build plugins
* Upgrade AssertJ
* Upgrade Spring to 7.x and enable integration tests for Java EA
* Adjust processor tests to use Spring with Jakarta Inject instead of Javax Inject since Spring 7 does not support Javax Inject
