+++ integrationtest/src/test/java/org/mapstruct/itest/tests/MavenIntegrationTest.java
@@ -82,7 +82,6 @@ public class MavenIntegrationTest {
 
     @ProcessorTest(baseDir = "jsr330Test")
     @EnabledForJreRange(min = JRE.JAVA_17)
-    @DisabledOnJre(JRE.OTHER)
     void jsr330Test() {
     }
 
@@ -176,7 +175,6 @@ public class MavenIntegrationTest {
 
     @ProcessorTest(baseDir = "springTest")
     @EnabledForJreRange(min = JRE.JAVA_17)
-    @DisabledOnJre(JRE.OTHER)
     void springTest() {
     }
