+++ processor/src/test/java/org/mapstruct/ap/test/decorator/jakarta/annotatewith/JakartaDecoratorAnnotateWithTest.java
@@ -5,6 +5,12 @@
  */
 package org.mapstruct.ap.test.decorator.jakarta.annotatewith;
 
+import java.util.Calendar;
+import jakarta.inject.Inject;
+import jakarta.inject.Named;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.mapstruct.ap.test.decorator.Address;
 import org.mapstruct.ap.test.decorator.AddressDto;
@@ -16,8 +22,13 @@ import org.mapstruct.ap.testutil.ProcessorTest;
 import org.mapstruct.ap.testutil.WithClasses;
 import org.mapstruct.ap.testutil.WithJakartaInject;
 import org.mapstruct.ap.testutil.runner.GeneratedSource;
+import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
 
 import static java.lang.System.lineSeparator;
+import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * Test for the application of @AnnotateWith on decorator classes with Jakarta component model.
@@ -32,11 +43,46 @@ import static java.lang.System.lineSeparator;
     TestAnnotation.class,
     JakartaAnnotateWithWithMapperDecorator.class
 })
+@ComponentScan(basePackageClasses = JakartaDecoratorAnnotateWithTest.class)
+@Configuration
 @WithJakartaInject
 public class JakartaDecoratorAnnotateWithTest {
 
     @RegisterExtension
     final GeneratedSource generatedSource = new GeneratedSource();
+
+    @Inject
+    @Named
+    private JakartaAnnotateWithMapper jakartaAnnotateWithMapper;
+
+    private ConfigurableApplicationContext context;
+
+    @BeforeEach
+    public void springUp() {
+        context = new AnnotationConfigApplicationContext( getClass() );
+        context.getAutowireCapableBeanFactory().autowireBean( this );
+    }
+
+    @AfterEach
+    public void springDown() {
+        if ( context != null ) {
+            context.close();
+        }
+    }
+
+    @ProcessorTest
+    public void shouldInvokeDecoratorMethods() {
+        Calendar birthday = Calendar.getInstance();
+        birthday.set( 1928, Calendar.MAY, 23 );
+        Person person = new Person( "Gary", "Crant", birthday.getTime(), new Address( "42 Ocean View Drive" ) );
+
+        PersonDto personDto = jakartaAnnotateWithMapper.personToPersonDto( person );
+
+        assertThat( personDto ).isNotNull();
+        assertThat( personDto.getName() ).isEqualTo( "Gary Crant" );
+        assertThat( personDto.getAddress() ).isNotNull();
+        assertThat( personDto.getAddress().getAddressLine() ).isEqualTo( "42 Ocean View Drive" );
+    }
 
     @ProcessorTest
     public void hasCorrectImports() {
