+++ processor/src/test/java/org/mapstruct/ap/test/injectionstrategy/jsr330/_default/Jsr330DefaultCompileOptionFieldMapperTest.java
@@ -5,13 +5,6 @@
  */
 package org.mapstruct.ap.test.injectionstrategy.jsr330._default;
 
-import javax.inject.Inject;
-import javax.inject.Named;
-
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.condition.DisabledOnJre;
-import org.junit.jupiter.api.condition.JRE;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.mapstruct.ap.test.injectionstrategy.shared.CustomerDto;
 import org.mapstruct.ap.test.injectionstrategy.shared.CustomerEntity;
@@ -21,13 +14,8 @@ import org.mapstruct.ap.testutil.ProcessorTest;
 import org.mapstruct.ap.testutil.WithClasses;
 import org.mapstruct.ap.testutil.WithJavaxInject;
 import org.mapstruct.ap.testutil.runner.GeneratedSource;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
 
 import static java.lang.System.lineSeparator;
-import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * Test field injection for component model jsr330.
@@ -43,48 +31,11 @@ import static org.assertj.core.api.Assertions.assertThat;
     CustomerJsr330DefaultCompileOptionFieldMapper.class,
     GenderJsr330DefaultCompileOptionFieldMapper.class
 })
-@ComponentScan(basePackageClasses = CustomerJsr330DefaultCompileOptionFieldMapper.class)
 @WithJavaxInject
-@Configuration
-@DisabledOnJre(JRE.OTHER)
 public class Jsr330DefaultCompileOptionFieldMapperTest {
 
     @RegisterExtension
     final GeneratedSource generatedSource = new GeneratedSource();
-
-    @Inject
-    @Named
-    private CustomerJsr330DefaultCompileOptionFieldMapper customerMapper;
-    private ConfigurableApplicationContext context;
-
-    @BeforeEach
-    public void springUp() {
-        context = new AnnotationConfigApplicationContext( getClass() );
-        context.getAutowireCapableBeanFactory().autowireBean( this );
-    }
-
-    @AfterEach
-    public void springDown() {
-        if ( context != null ) {
-            context.close();
-        }
-    }
-
-    @ProcessorTest
-    public void shouldConvertToTarget() {
-        // given
-        CustomerEntity customerEntity = new CustomerEntity();
-        customerEntity.setName( "Samuel" );
-        customerEntity.setGender( Gender.MALE );
-
-        // when
-        CustomerDto customerDto = customerMapper.asTarget( customerEntity );
-
-        // then
-        assertThat( customerDto ).isNotNull();
-        assertThat( customerDto.getName() ).isEqualTo( "Samuel" );
-        assertThat( customerDto.getGender() ).isEqualTo( GenderDto.M );
-    }
 
     @ProcessorTest
     public void shouldHaveFieldInjection() {
