+++ processor/src/test/java/org/mapstruct/ap/test/injectionstrategy/jsr330/compileoptionconstructor/Jsr330CompileOptionConstructorMapperTest.java
@@ -5,10 +5,6 @@
  */
 package org.mapstruct.ap.test.injectionstrategy.jsr330.compileoptionconstructor;
 
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.condition.DisabledOnJre;
-import org.junit.jupiter.api.condition.JRE;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.mapstruct.ap.test.injectionstrategy.shared.CustomerDto;
 import org.mapstruct.ap.test.injectionstrategy.shared.CustomerEntity;
@@ -19,14 +15,8 @@ import org.mapstruct.ap.testutil.WithClasses;
 import org.mapstruct.ap.testutil.WithJavaxInject;
 import org.mapstruct.ap.testutil.compilation.annotation.ProcessorOption;
 import org.mapstruct.ap.testutil.runner.GeneratedSource;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
 
 import static java.lang.System.lineSeparator;
-import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * Test constructor injection for component model jsr330 with compile option
@@ -43,47 +33,11 @@ import static org.assertj.core.api.Assertions.assertThat;
     GenderJsr330CompileOptionConstructorMapper.class
 })
 @ProcessorOption( name = "mapstruct.defaultInjectionStrategy", value = "constructor")
-@ComponentScan(basePackageClasses = CustomerJsr330CompileOptionConstructorMapper.class)
-@Configuration
 @WithJavaxInject
-@DisabledOnJre(JRE.OTHER)
 public class Jsr330CompileOptionConstructorMapperTest {
 
     @RegisterExtension
     final GeneratedSource generatedSource = new GeneratedSource();
-
-    @Autowired
-    private CustomerJsr330CompileOptionConstructorMapper customerMapper;
-    private ConfigurableApplicationContext context;
-
-    @BeforeEach
-    public void springUp() {
-        context = new AnnotationConfigApplicationContext( getClass() );
-        context.getAutowireCapableBeanFactory().autowireBean( this );
-    }
-
-    @AfterEach
-    public void springDown() {
-        if ( context != null ) {
-            context.close();
-        }
-    }
-
-    @ProcessorTest
-    public void shouldConvertToTarget() {
-        // given
-        CustomerEntity customerEntity = new CustomerEntity();
-        customerEntity.setName( "Samuel" );
-        customerEntity.setGender( Gender.MALE );
-
-        // when
-        CustomerDto customerDto = customerMapper.asTarget( customerEntity );
-
-        // then
-        assertThat( customerDto ).isNotNull();
-        assertThat( customerDto.getName() ).isEqualTo( "Samuel" );
-        assertThat( customerDto.getGender() ).isEqualTo( GenderDto.M );
-    }
 
     @ProcessorTest
     public void shouldHaveConstructorInjectionFromCompileOption() {
